Add macro OFFUNISKIP
This means use official Unicode code point numbering, not native.  Doing
this converts the existing UNISKIP calls in the code to refer to native
code points, which is what they meant anyway.  The terminology is
somewhat ambiguous, but I don't think will cause real confusion.
NATIVE_SKIP is also introduced for situations where it is important to
be precise.