pp.c, regexec.c: Declare buffers large enough
These three buffers are not declared with the proper size.  There is
a #define to use in these declarations, so use it.  These matter only on
EBCDIC platforms, where the one in pp.c prior to this commit could cause
a buffer overrun there.

The others shouldn't because what is being used is known (smaller) size.